packaging: spec: split ovirt-host-deploy-offline into own spec
rpm dependencies are quite primitive and set at build time and not at
install time, the dmidecode dependency cannot be use as it is platform
specific in noarch package.

rpm does not allow building base package as noarch while subpackages
as arch specific, we cannot have the offline subpackage depend on
arch specific packages.

the reasonable solutions are:

1. make the entire package arch specific, it is illogical to accept rpm
   limitation and enforce it on good package.

2. create a package that provides dmidecode, so it can be installed at
   archs that do not provide dmidecode. this again is ugly.

3. move ovirt-host-deploy-offline into own arch specific primary
   package.

solution 3 was implemented.

a target of offline-tarball was added to create a tarball that can be
served as source for packaging.

Change-Id: I26513d02fd1c95a2a515781884a7283e77a5c2ff
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>