hosted-engine: prevent a race cond. between ha-agent and vdsm
Without this, host-deploy will first start ovirt-ha-agent that will
bring up vdsm due to systemd dependencies but, just after that,
host-deploy will (re)start vdsm.
If ovirt-ha-agent is not quick enough, vdsm will be restarted in
the middle of connectStorageServer that will not complete,
because vdsm was stopped and does not continue such operations
upon restart.
On the other side, the timeout waiting for connectStorageServer
to be completed is pretty long (900 seconds) so ovirt-ha-agent
will take a lot of time to realize the issue.
In the mean time, ovirt-ha-broker will fail accessing the shared
storage and vdsm get the error and reports back a bad status
to the engine and so adding the host from the webui could fail
if ovirt-ha-agent is not fast enough.
Reversing the start order of vdsm and ovirt-ha-agent
will prevent this since ovirt-ha-agnet has Wants=vdsmd.service and
After=vdsmd.service and both are met.

Change-Id: I2e5dd0cb3c6cbfec00286b063558d379ade1def4
Bug-Url: https://bugzilla.redhat.com/1349532
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>