kernel: allow installation of RT kernel+kvm
Realtime kernel and KVM are separate packages. Up until now, we had
no way of installing them - and there really wasn't a need for it. As
we are working to add RT operation support, we should also be able to
handle the installation.

The required packages are kernel-rt and kernel-rt-kvm. Installing
these packages does is necessary but insufficient step in enabling RT
host.

Change-Id: I7c9b5abf6697e8653d8ab85094366555b302ff93
Bug-Url: https://bugzilla.redhat.com/1425447
Signed-off-by: Martin Polednik <mpolednik@redhat.com>