set 'virtual-host' as the default tuned profile
'virtual-host' is the recommanded tuned profile for hypervisors.
Same profile is recommanded for hyperconverged installation with
gluster. So when there is no tuned profile defined at the cluster
level then 'virtual-host' will be used if virt service is enabled.

Note: If there is no profile defined at the cluster level and only
gluster service is enabled not virt service then we will not
set any tuned profile.

Change-Id: Ifab913bb55dc4708fe33799459a83c6ea7d43ede
Bug-Url: https://bugzilla.redhat.com/1380739
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>