vdsm: restart imageio daemon on vdsm start
imageio-daemon loads vdsm's PKI for SSL communication
on service start. Hence, when engine enrolls a new certificate,
the daemon must be restarted in order to reload it.

So the host-deploy should make sure to restart the daemon upon
a successful deployment. This is also relevant for hosted-engine
setup flow, in which the engine enrolls the certificate to the
host after the daemon is already started.

Change-Id: I3a998a4901fe7c91da8d0d904762dc57eac0b6de
Bug-Url: https://bugzilla.redhat.com/1411491
Signed-off-by: Daniel Erez <derez@redhat.com>