tune/iosched: Keep default
Do not configure default ioscheduler anymore. We used to configure it
to 'deadline', which is the default for supported kernels, thus not
needed anymore.

Remove old configuration file if found.

With the previous configuration we wrote, it was impossible to override
host-deploy's change by setting an attribute in a later conf file, which
is what some storage vendors recommend to do. With this patch this
should be trivial.

Change-Id: I2e98afaed5444efd67cc5f7ad3a3b40aa1f76933
Bug-Url: https://bugzilla.redhat.com/1381219
Signed-off-by: Yedidyah Bar David <didi@redhat.com>
(cherry picked from commit 8c42c3f4150000d5f9afd36fb0b4b9770edcb818)