packaging: consume ovirt-host package
Instead of requiring individual mandatory packages,
just require ovirt-host package.
It is possible to require a minimal version to be
installed on the host using environment configuration.

This requires ovirt-host package to be backported to ovirt-4.1,
ovirt-4.0 and ovirt-3.6 in order to be able to deploy hosts with
such version.

Bug-Url: https://bugzilla.redhat.com/1461081
Change-Id: I4b1d45d61cee202737503070dd273178822dfd53
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>