packaging: make qemu-kvm-tools optional
qemu-kvm-tools disappeared from Fedora 25 and it
was a nice way to discover it's there since 2012
but unused right now.
We stopped requiring it in 4.2 introducing a regression
when trying to deploy oVirt Node 4.2 with oVirt Engine 4.1
since qemu-kvm-tools is not included in Node.

This patch make it optional, keeping it if available in order
to not change existing 4.1 behavior but not failing on its
absence for 4.2 compatibility.

Change-Id: I09e73a241c341ca7f2b6bba4846c2e36cd66f7a9
Bug-Url: https://bugzilla.redhat.com/1501761
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>