kernel: fips: add boot parameter if fips=1 is there
When the user choose to enforce FIPS kernel
parameter from the engine, the engine is simply
going to add fips=1 but this is going to prevent
the host from rebooting if /boot is not wrote on the root
partition but on a separate one.
Detect 'fips=1' without 'boot=', identify the boot
partition uuid and add
boot=UUID=<boot_p_uuid> to the kernel parameters.

Change-Id: I5edb0b0c632ac3b4cd45a868e608245ee8d070ef
Bug-Url: https://bugzilla.redhat.com/1737926
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
(cherry picked from commit f28a429aef0411787e08f0c8a5705cee54361e01)